+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.7.10 (unreleased)
 
 - move `Bundle`, `AppBundle`, and `BundleBlueprint` into the `flask_unchained.bundles` module
+- set up automatic dependency injection on commands
 
 ## v0.7.9 (2019/05/19)